+++ XUIKit/XUIKit/XUITextField.m
@@ -91,7 +91,7 @@
 }
 
 - (void)setLeftViewMode:(XUITextFieldViewMode)leftViewMode{
-    _textfieldFlags.leftViewMode = leftViewMode;
+    _textfieldFlags.leftViewMode = (unsigned int)leftViewMode;
     [self setNeedsDisplay:YES];
 }
 
@@ -100,7 +100,7 @@
 }
 
 - (void)setRightViewMode:(XUITextFieldViewMode)rightViewMode{
-    _textfieldFlags.rightViewMode = rightViewMode;
+    _textfieldFlags.rightViewMode = (unsigned int)rightViewMode;
     [self setNeedsDisplay:YES];
 }
 
@@ -109,7 +109,7 @@
 }
 
 - (void)setClearButtonMode:(XUITextFieldViewMode)clearButtonMode{
-    _textfieldFlags.clearButtonMode = clearButtonMode;
+    _textfieldFlags.clearButtonMode = (unsigned int)clearButtonMode;
     [self setNeedsDisplay:YES];
 }
